src/main/java/com/duocuc/products_srv/dto/ProductIemDto.java
package com.duocuc.products_srv.dto;

import java.util.Objects;

public class ProductIemDto {

  private Long id;
  private String name;
  private Double price;

  public ProductIemDto(Long id, String name, Double price) {
    this.id = id;
    this.name = name;
    this.price = price;
  }

  // Getters y setters
  public Long getId() {
    return id;
  }

  public void setId(Long id) {
    this.id = id;
  }

  public String getName() {
    return name;
  }

  public void setName(String name) {
    this.name = name;
  }

  public Double getPrice() {
    return price;
  }

  public void setPrice(Double price) {
    this.price = price;
  }

  @Override
  public boolean equals(Object o) {
    if (this == o)
      return true;
    if (o == null || getClass() != o.getClass())
      return false;
    ProductIemDto that = (ProductIemDto) o;
    return Objects.equals(id, that.id) &&
        Objects.equals(name, that.name) &&
        Objects.equals(price, that.price);
  }

  @Override
  public int hashCode() {
    return Objects.hash(id, name, price);
  }

  @Override
  public String toString() {
    return "ProductIemDto{" +
        "id=" + id +
        ", name='" + name + '\'' +
        ", price=" + price +
        '}';
  }
}
